コントローラーをv1, v2で分離
