+++ app/controllers/user/UserControllerV2.scala
@@ -1,40 +1,34 @@
 package controllers.user
 
-import javax.inject.Inject
-import domain.{User, UserUpdateRequest}
-import infra.dbclients.{UserDBClient, UserDBClientV2}
+import controllers.user.RequestConverter.{
+  userReads,
+  userUpdateRequestReads,
+  userWrites
+}
+import controllers.user.vo.UserUpdateRequest
+import domain.User
+import infra.dbclients.v1.UserDBClient
+import infra.dbclients.v2.UserDBClientV2
 import play.api.Logging
 import play.api.libs.json.Json.toJson
 import play.api.mvc.{BaseController, ControllerComponents}
 
-import scala.concurrent.Future
-import scala.concurrent.ExecutionContext.Implicits.global
+import javax.inject.Inject
 import scala.Function.const
+import scala.concurrent.ExecutionContext.Implicits.global
+import scala.concurrent.Future
 
 /** controller for user endpoint
+  *
   * @param dbClient user db client
   */
-class UserController @Inject() (
+class UserControllerV2 @Inject() (
     dbClient: UserDBClient,
     dbClientV2: UserDBClientV2
 )(
     val controllerComponents: ControllerComponents
 ) extends BaseController
     with Logging {
-
-  import RequestConverter._ // -- (1)
-
-  /** action for creating user
-    * @return
-    */
-  def post = Action.async(parse.json) { req =>
-    req.body
-      .validate[User]
-      .fold( // -- (2)
-        invalid => Future(BadRequest), // -- (3)
-        user => dbClient.createIfNotExist(user).map(const(Ok)) // -- (4)
-      )
-  }
 
   def list = Action.async {
     logger.info("start list")
@@ -57,7 +51,7 @@ class UserController @Inject() (
   }
 
   // fixme 何回もpostできてしまうのを直す
-  def postV2 = Action.async(parse.json) { req =>
+  def post = Action.async(parse.json) { req =>
     logger.info("start post")
     req.body
       .validate[User]
