+++ app/modules/AppDependencyModule.scala
@@ -3,7 +3,8 @@ package modules
 import com.amazonaws.services.dynamodbv2.document.DynamoDB
 import com.google.inject.AbstractModule
 import controllers.user.UserController
-import infra.dbclients.{DynamoDBProvider, DynamoDBProviderV2, UserDBClient}
+import infra.dbclients.v1.{DynamoDBProvider, UserDBClient}
+import infra.dbclients.v2.DynamoDBProviderV2
 import software.amazon.awssdk.services.dynamodb.DynamoDbAsyncClient
 
 /** resolves DI
